Put back the call to the internal class for Windows users.  A better solution is needed for linux.
I need a way to load a DLL into an AppDomain that is not my own.
I don't think a AppDomainManager would work and you need .NET v4 to
link to one from a config file anyways.  And I can't figure out
how to get at an instance of AppDomain that is not my own.  Finding some
object in the .NET framework that has will take a delegate and then
run it in the context of the hosting environment would work, but I have
not been able to find a class that would work yet.